handle repeated START condition correctly
The Atmel TWI interface can indicate an "unexpected STOP" when
there's a repeated START condition. Don't try to treat it as a bus
error, and don't wait for TWSTO to clear even if we do have a bus
error (and it's not officially documented that reading TWSTO is
meaningful).

Previously, a repeated START condition would cause the receiver
to loop waiting for TWSTO to clear. This meant it would miss
an address match in a repeated START condition, resulting in a
default NAK until the next START condition (which might clear
TWSTO?).

Signed-off-by: Taylor Yu <tlyu@mit.edu>